requirements.txt
Flask==2.2.2
flask_cors==3.0.10
numpy==1.23.4
Pillow==9.3.0
tensorflow==2.11.0
opencv-python==4.6.0.66
gunicorn